+++ nosi/nosi/flash_cache_engine/flash_h2d_mask_bias.py
@@ -24,9 +24,10 @@ def flash_h2d_from_mask_bias_kernel(
     D: tl.constexpr,
 ):
 
-    pid_b = tl.program_id(0)
-    pid_h = tl.program_id(1)
-    pid_m = tl.program_id(2)
+    # The same wide offsets cover KV and bias, on both sides of the copy.
+    pid_b = tl.program_id(0).to(tl.int64)
+    pid_h = tl.program_id(1).to(tl.int64)
+    pid_m = tl.program_id(2).to(tl.int64)
 
     # load offset
     cpu_block_id = tl.load(
